Move draw leaf function into separate js file
